Add validations

+++ src/infra/sqlalchemy/repos/product.py
@@ -46,38 +46,45 @@ class RepoProduct():
 
 
     def update(self, product: schemas.Product):
-        print(product)
+        db_product = self.db.query(models.Product).filter(models.Product.id == product.id).first() is not None
 
-        update_stmt = update(models.Product).where(
-            models.Product.id == product.id
-        ).values(
-            sku=product.sku,
-            name=product.name,
-            description=product.description,
-            price=product.price,
-            category_id=product.category_id,
-            size_weight=product.size_weight,
-            size_width=product.size_width,
-            size_height=product.size_height,
-            size_length=product.size_length,
-            qty_items_per_box=product.qty_items_per_box,
-            ean=product.ean,
-            promo=product.promo,
-            promo_discount=product.promo_discount,
-            qty_stock=product.qty_stock
-        )
+        if db_product:
+            update_stmt = update(models.Product).where(
+                models.Product.id == product.id
+            ).values(
+                sku=product.sku,
+                name=product.name,
+                description=product.description,
+                price=product.price,
+                category_id=product.category_id,
+                size_weight=product.size_weight,
+                size_width=product.size_width,
+                size_height=product.size_height,
+                size_length=product.size_length,
+                qty_items_per_box=product.qty_items_per_box,
+                ean=product.ean,
+                promo=product.promo,
+                promo_discount=product.promo_discount,
+                qty_stock=product.qty_stock
+            )
 
-        self.db.execute(update_stmt)
-        self.db.commit()
+            self.db.execute(update_stmt)
+            self.db.commit()
 
-        return  product
+            return True
+        raise ValueError('ID do produto não encontrado.')
     
 
     def delete(self, product_id: int):
+        product_associated = self.db.query(models.OrderItem).filter(models.OrderItem.product_id == product_id).first()
+        
+        if product_associated:
+            raise ValueError('Não é possível excluir o produto, pois está associado a um pedido.')
+        
         db_product = self.db.query(models.Product).filter(models.Product.id == product_id).first()
         
         if db_product:
             self.db.delete(db_product)
             self.db.commit()
             return True
-        return False
+        raise ValueError('ID do produto não encontrado.')
